change: Change to stateful components
This allows each component to handle its state. I can now render
components based on a list and the page will know (or, more
accurately, not need to know) what needs to be updated when. The
components themselves take care of updates (at a preset interval,
determined by the view).