feat: reduce load on API servers; use data stores for components
Instead of having each component fetch its own data from the API on
every update (which means multiple fetches could occur on a single
page for the same information if a component is rendered more than
once), a dynamically-generated GenServer is created that handles the
fetching and data. When the GenServer is generated dynamically, it
uses a computed name -- the component name and the user id. This means
that every component with that name rendered by that user will share
the same data. Furthermore, this means that an infinite number of
components could be rendered on an infinite number of dashboards and
as long as the same logged-in user was looking at the dashboard, it
would make 1 total api call to Planning Center.

Not only does this greatly reduce the load on Planning Center's API,
it reduces the load we have to do to get that information and
distribute it across all the instances of the same component.