feat: allow api paths to be configurable
Using Dashboard ComponentConfigurations, an api_path can have
configurable paths. It handles this by using `String.replace/3` on the
api_path in the database for a configuration. For example, the list
results component takes a list_id as a configuration. That list_id is
needed to form the correct api_path.