change: terminate Component GenServers if no one's subscribed
If no one is subscribed to a Component GenServer, it won't fetch the
api endpoint and it also won't update the `last_update` timestamp. If
there are no subscribers, and it's been more that 10 minutes since the
last listened-to update, terminate the GenServer. We can easily spin
it back up again when needed.