fix: don't attempt to resequence if sequence is the same
If a user picked up the first component and dropped it to the left, it
would resequence the component to have the sequence of -1. This is
obviously not what we want and produced bugs. Instead, make sure that
the new sequence doesn't equal the old before resequencing. Do this in
three places: one in the view's javascript event, one closer to the
view, and one closer to the database.